Module Univ_map

module type S = sig ... end

module type S1 = sig ... end

module type Key = sig ... end

module type Data = sig ... end

module Type_id_key : Key with type 'a t = 'a Base.Type_equal.Id.t

include S with type 'a data = 'a and module Key := Type_id_key

type t

val sexp_of_t : t -> Sexplib0.Sexp.t

type 'a data = 'a

include Base.Invariant.S with type t := t

val invariant : t -> unit

val empty : t

val singleton : 'a Type_id_key.t -> 'a data -> t

val is_empty : t -> Base.bool

val set : t -> key:'a Type_id_key.t -> data:'a data -> t

val mem : t -> 'a Type_id_key.t -> Base.bool

val mem_by_id : t -> Base.Type_equal.Id.Uid.t -> Base.bool

val find : t -> 'a Type_id_key.t -> 'a data Base.option

val find_exn : t -> 'a Type_id_key.t -> 'a data

val add : t -> key:'a Type_id_key.t -> data:'a data -> [ `Ok of t | `Duplicate ]

val add_exn : t -> key:'a Type_id_key.t -> data:'a data -> t

val change : t -> 'a Type_id_key.t -> f:('a data Base.option -> 'a data Base.option) -> t

val change_exn : t -> 'a Type_id_key.t -> f:('a data -> 'a data) -> t

val update : t -> 'a Type_id_key.t -> f:('a data Base.option -> 'a data) -> t

val remove : t -> 'a Type_id_key.t -> t

val remove_by_id : t -> Base.Type_equal.Id.Uid.t -> t

module Packed : sig ... end

val key_id_set : t -> Base.Set.M(Base.Type_equal.Id.Uid).t

val to_alist : t -> Packed.t Base.list

to_alist t returns all values in t, in increasing order of key type-id name.

val of_alist_exn : Packed.t Base.list -> t

val find_packed_by_id : t -> Base.Type_equal.Id.Uid.t -> Packed.t Base.option

val find_packed_by_id_exn : t -> Base.Type_equal.Id.Uid.t -> Packed.t

val type_equal : (t, Packed.t Base.Map.M(Base.Type_equal.Id.Uid).t) Base.Type_equal.t

module Key = Base.Type_equal.Id

This binding is convenient because existing call sites often refer to Univ_map.Key.create.

module Make (Key : Key) (Data : Data) : S with type 'a data = 'a Data.t and module Key = Key

module Make1 (Key : Key) (Data : sig ... end) : S1 with type ('s, 'a) data = ('s, 'a) Data.t and module Key = Key

module Merge (Key : Key) (Input1_data : Data) (Input2_data : Data) (Output_data : Data) : sig ... end

module Merge1 (Key : Key) (Input1_data : sig ... end) (Input2_data : sig ... end) (Output_data : sig ... end) : sig ... end

module With_default : sig ... end

keys with associated default values, so that find is no longer partial

module With_fold : sig ... end

keys that map to an accumulator value with an associated fold operation

module Multi : sig ... end

list-accumulating keys with a default value of the empty list